Move erased stuff into seperate modules
