+++ packages/ec-core/src/operator/mutator/mod.rs
@@ -2,6 +2,11 @@ use rand::Rng;
 
 use super::{Composable, Operator};
 
+#[cfg(feature = "erased")]
+mod erased;
+#[cfg(feature = "erased")]
+pub use erased::*;
+
 /// Mutate a genome of type `G` generating a new
 /// genome of the same type (`G`).
 ///
@@ -63,39 +68,6 @@ pub trait Mutator<G> {
     /// genome. This will usually be because the given `genome` is invalid in
     /// some way, thus making the mutation impossible.
     fn mutate<R: Rng + ?Sized>(&self, genome: G, rng: &mut R) -> Result<G, Self::Error>;
-}
-
-#[cfg(feature = "erased")]
-pub trait DynMutator<G, E = Box<dyn std::error::Error + Send + Sync>> {
-    /// # Errors
-    ///
-    /// This will return an error if there is an error mutating the given
-    /// genome. This will usually be because the given `genome` is invalid in
-    /// some way, thus making the mutation impossible.
-    fn dyn_mutate(&self, genome: G, rng: &mut dyn rand::RngCore) -> Result<G, E>;
-}
-
-#[cfg(feature = "erased")]
-static_assertions::assert_obj_safe!(DynMutator<()>);
-
-#[cfg(feature = "erased")]
-impl<T, G, E> DynMutator<G, E> for T
-where
-    T: Mutator<G, Error: Into<E>>,
-{
-    fn dyn_mutate(&self, genome: G, rng: &mut dyn rand::RngCore) -> Result<G, E> {
-        self.mutate(genome, rng).map_err(Into::into)
-    }
-}
-
-#[cfg(feature = "erased")]
-#[ec_macros::dyn_ref_impls]
-impl<G, E> Mutator<G> for &dyn DynMutator<G, E> {
-    type Error = E;
-
-    fn mutate<R: Rng + ?Sized>(&self, genome: G, mut rng: &mut R) -> Result<G, Self::Error> {
-        (**self).dyn_mutate(genome, &mut rng)
-    }
 }
 
 /// A wrapper that converts a [`Mutator`] into an [`Operator`].
